pyproject.toml
[project]
name = "invoice-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datetime>=5.5",
    "jinja2>=3.1.4",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.3",
    "weasyprint>=62.3",
]
